Started Matrix/Vector expression templates
	new file:   MVExpr.h
	renamed:    VectorN.h -> MatVec/VectorN.h
	new file:   MatrixNM.h
	modified:   Vector2.h
	modified:   Vector3.h
	modified:   Vector4.h
	modified:   Vector6.h
	modified:   VectorN.h
	modified:   ../../tests/CMakeLists.txt
	new file:   ../../tests/matLUP/CMakeLists.txt
	renamed:    ../../tests/matvec/main.cpp -> ../../tests/matLUP/main.cpp
	modified:   ../../tests/matvec/main.cpp

+++ cvt/math/Vector2.h
@@ -60,7 +60,7 @@ namespace cvt {
         T			normalize( void );		// returns length
         void		clamp( const Vector2<T> &min, const Vector2<T> &max );
 
-        int			dimension( void ) const;
+        size_t		dimension( void ) const;
 
         bool        isEqual( const Vector2<T> & other, T epsilon ) const;
 
@@ -311,7 +311,7 @@ namespace cvt {
     }
 
     template<typename T>
-    int Vector2<T>::dimension( void ) const
+    size_t Vector2<T>::dimension( void ) const
     {
         return 2;
     }
